+++ crates/ic-dbms-api/src/dbms/table.rs
@@ -8,8 +8,10 @@ use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
-pub use self::column_def::{ColumnDef, ForeignKeyDef};
-pub use self::record::{InsertRecord, TableColumns, TableRecord, UpdateRecord, ValuesSource};
+pub use self::column_def::{CandidColumnDef, CandidForeignKeyDef, ColumnDef, ForeignKeyDef};
+pub use self::record::{
+    InsertRecord, TableColumns, TableRecord, UpdateRecord, ValuesSource, flatten_table_columns,
+};
 pub use self::schema::{TableFingerprint, TableSchema};
 
 /// Table related errors